+++ todo_app/data/item.py
@@ -8,13 +8,7 @@ class Item:
     def from_trello_card(card, list): 
         return Item(card['id'], card['name'], list['name'])
 
-class ViewModel:
-	def __init__(self, items):
-		self._items = items
- 
-	@property
-	def items(self):
-		return self._items
+
