+++ todo_app/data/view_model.py
@@ -1,0 +1,12 @@
+class ViewModel:
+	def __init__(self, items):
+		self._items = items
+ 
+	@property
+	def items(self):
+		return self._items
+		
+	@property
+	def todo_items(self):
+		return []
+
