--- todo_app/test_view_status.py
@@ -1,15 +1,0 @@
-import pytest
-from todo_app.data.item import View_model
-
-#Arrenge -  Create an instance of ViewModel containing some example Item objects with various statuses
-
-def view_test():
-    status_1 = 'Doing'
-    status_2 = 'Done'
-
-
-#Act -  Get the result of the view model's new doing_items property (though this will actually be an empty list for now).
-
-
-#Assert - Write one or more assert statements checking the result of the "Act" step equals what you want
-
